requirements.txt
easyocr==1.7.2
filelock==3.29.0
fsspec==2026.4.0
ImageIO==2.37.3
Jinja2==3.1.6
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.2
pillow==12.2.0
pyclipper==1.4.0
python-bidi==0.6.7
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
sympy==1.14.0
tifffile==2026.5.2
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
